JSUI-2417 MLFacet public reference documentation (#1046)
* Add configuration for MLFacet playground example

https://coveord.atlassian.net/browse/JSUI-2417

* Document FacetSortCriteria enum

https://coveord.atlassian.net/browse/JSUI-2417

* Document MLFacet component

https://coveord.atlassian.net/browse/JSUI-2417

* Fix IMLFacetOptions import in unit tests

https://coveord.atlassian.net/browse/JSUI-2417

* Capitalize enum values

https://coveord.atlassian.net/browse/JSUI-2417

* Remove formatting for enum values

https://coveord.atlassian.net/browse/JSUI-2417

* Revert enum casing refactor

https://coveord.atlassian.net/browse/JSUI-2417

* Remove incomplete sentence from MLFacet description

https://coveord.atlassian.net/browse/JSUI-2417
