+++ unitTests/ui/MLFacet/MLFacetValues/MLFacetValueTest.ts
@@ -1,8 +1,7 @@
 import * as Globalize from 'globalize';
 import { MLFacetValue } from '../../../../src/ui/MLFacet/MLFacetValues/MLFacetValue';
-import { IMLFacetOptions } from '../../../../src/ui/MLFacet/MLFacetOptions';
 import { MLFacetTestUtils } from '../MLFacetTestUtils';
-import { MLFacet } from '../../../../src/ui/MLFacet/MLFacet';
+import { MLFacet, IMLFacetOptions } from '../../../../src/ui/MLFacet/MLFacet';
 import { FacetValueState } from '../../../../src/rest/Facet/FacetValueState';
 
 export function MLFacetValueTest() {
